flower list & design

+++ 00003951_DBSD_CW2/DataAccess/FlowerManager.cs
@@ -99,7 +99,7 @@ namespace _00003951_DBSD_CW2.DataAccess
                                 Name = reader.GetString(1),
                                 Description = reader.GetString(2),
                                 ImgUrl = reader.GetString(3),
-                                Price = reader.GetFloat(4),
+                                Price = reader.GetDouble(4),
                                 Remaining = reader.GetInt32(5),
                                 FlowerCategoryId = reader.GetInt32(6)
                             };
